app/back/dit-app/requirements.txt
# GENERAL
# python-dotenv==1.0.1
requests==2.32.3

# POSTGRES
psycopg2-binary==2.9.10

# DJANGO/DRF WITH POSTGIS
Django==4.2.17
djangorestframework==3.15.0
djangorestframework-gis==1.1